Implemented the notifications function: used a presenter to handle the views requests and retrieve the notifications from the database.
Added the presenter to the ContentHolder: request is sent when we click on the sideBar-notification-tab, which is a part of ContentHolder.

The notificationEntity relationship to the other entities is now temporarily unidirectional. Bidirectional caused some issues, as some parts are still missing in the database(extra research needed).
TL;DR removed the OneToOne side on the notifications.